Overhauled trailer logic to allow for trailers to request hookups.  Also allows for multiple trailers.  Allows for grouped hitches.  Allows for defining if trailers can attach other trailers from the truck towing them.  Allows definable hitch distances per #775.

+++ src/main/java/minecrafttransportsimulator/entities/instances/AEntityVehicleB_Rideable.java
@@ -202,7 +202,7 @@ abstract class AEntityVehicleB_Rideable extends AEntityE_Multipart<JSONVehicle>{
 	 * Calculates the current mass of the vehicle.
 	 * Includes core mass, player weight, and cargo.
 	 */
-	protected float getCurrentMass(){
+	protected int getCurrentMass(){
 		int currentMass = definition.motorized.emptyMass + 100*locationRiderMap.values().size();
 		for(APart part : parts){
 			if(part instanceof PartInteractable){
